+++ libcnb/src/detect.rs
@@ -2,7 +2,7 @@
 
 use crate::buildpack::Buildpack;
 use crate::data::buildpack::StackId;
-use crate::{data::build_plan::BuildPlan, data::buildpack::SingleBuildpackDescriptor};
+use crate::{data::build_plan::BuildPlan, data::buildpack::ComponentBuildpackDescriptor};
 use std::fmt::Debug;
 use std::path::PathBuf;
 
@@ -12,7 +12,7 @@ pub struct DetectContext<B: Buildpack + ?Sized> {
     pub buildpack_dir: PathBuf,
     pub stack_id: StackId,
     pub platform: B::Platform,
-    pub buildpack_descriptor: SingleBuildpackDescriptor<B::Metadata>,
+    pub buildpack_descriptor: ComponentBuildpackDescriptor<B::Metadata>,
 }
 
 /// Describes the result of the detect phase.